% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calculate_laplace_entropy}
\alias{calculate_laplace_entropy}
\title{Calculate Laplace entropy for a vector of transcript-level
expression values of one gene.}
\usage{
calculate_laplace_entropy(x, norm = TRUE)
}
\arguments{
\item{x}{Vector of expression values.}

\item{norm}{If \code{TRUE}, the entropy values are normalized to the number
of transcripts for each gene. The normalized entropy values are always
between 0 and 1. If \code{FALSE}, genes cannot be compared to each other,
due to possibly different maximum entropy values.}
}
\value{
A single gene-level Laplace entropy value.
}
\description{
Calculate Laplace entropy for a vector of transcript-level
expression values of one gene.
}
\details{
The function calculates a Laplace entropy value as part of different
diversity calculations. Given a vector of transcript-level expression values
of a gene, this function characterize the diversity of splicing isoforms for a
gene. If there only one single transcript, the resulted index will be NaN, as
diversity cannot be calculated. If the expression of the given gene is 0, the
diversity index will be NA.
}
