% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calculate_simpson}
\alias{calculate_simpson}
\title{#' Calculate Simpson index for a vector of transcript-level
expression values of one gene.}
\usage{
calculate_simpson(x)
}
\arguments{
\item{x}{Vector of expression values.}
}
\value{
A single gene-level Simpson index.
}
\description{
#' Calculate Simpson index for a vector of transcript-level
expression values of one gene.
}
\details{
The function calculates a Simpson index as part of different
diversity calculations. Given a vector of transcript-level expression values
of a gene, this function characterize the diversity of splicing isoforms for a
gene. If there only one single transcript, the resulted index will be NaN, as
diversity cannot be calculated. If the expression of the given gene is 0, the
diversity index will be NA.
}
