% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{calculate_tsallis_entropy}
\alias{calculate_tsallis_entropy}
\title{Calculate Tsallis entropy for a vector of transcript-level
expression values of one gene.}
\usage{
calculate_tsallis_entropy(x, q = 2, norm = TRUE)
}
\arguments{
\item{x}{Vector of expression values.}

\item{q}{Tsallis entropy parameter (q > 0). Can be a single value or a numeric vector. Default is 2. If q = 1, the function returns the Shannon entropy.}

\item{norm}{If \code{TRUE}, the entropy values are normalized to the number
of transcripts for each gene. The normalized entropy values are always
between 0 and 1. If \code{FALSE}, genes cannot be compared to each other,
due to possibly different maximum entropy values.}
}
\value{
A single gene-level Tsallis entropy value.
}
\description{
Calculate Tsallis entropy for a vector of transcript-level
expression values of one gene.
}
\details{
The function calculates the Tsallis entropy, a generalization of Shannon entropy.
For q → 1, Tsallis entropy converges to Shannon entropy.
If there is only a single transcript, the value will be NaN.
If the expression of the given gene is 0, the value will be NA.
}
\examples{
x <- rnbinom(5, size = 10, prob = 0.4)
tsallis <- calculate_tsallis_entropy(x, q = 2)

}
