% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{TCGA Luminal A breast cancer dataset}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://portal.gdc.cancer.gov/legacy-archive}{TCGA Legacy}
}
\usage{
data(example_dataset)
}
\description{
Data from The Cancer Genome Atlas, downloaded on 12th April, 2020. It
contains transcript level read counts of 20 patient with Luminal A type
breast cancer (primary tumor and solid normal samples).
}
\references{
The Cancer Genome Atlas Network (2012) Nature 490, 61–70
  (\href{https://doi.org/10.1038/nature11412}{doi:10.1038/nature11412})
}
\keyword{datasets}
